refactor: split hooks.spec.ts into smaller test files
